refactor(libraryTest): 주석 가독성을 위해 변경
